Initial commit of Zero Reboot Installer
The scheme described in the README.md here is something I've been
thinking about for many years.

In October of 2024 I manually prepared a USB drive for this purpose,
documenting none of the process.

I created the install.sh you see here, which I used in a Lambda Show and
Tell meeting to demo a PoC of this concept.

The install.sh was in /etc/skel/Desktop/install.sh so that it would be
visible on the desktop of the newly created user.

It's now December 2025, and I'm still using the system that I installed
in that Show and Tell, and I'd like to make it at least a proof of
concept that's reproducible / doesn't require manual steps.

So, I created a new directory, changed to it, ran `git init .` then `cp ~/Desktop/install.sh .`, wrote
the README.md, and that's what this commit contains.

Clearly, there's a lot more work yet to do.